requirements.txt
assertpy==1.1
certifi==2025.1.31
charset-normalizer==3.4.1
Faker==35.2.0
greenlet==3.1.1
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.5
MarkupSafe==3.0.2
numpy==2.2.2
packaging==24.2
pandas==2.2.3
playwright==1.50.0
pluggy==1.5.0
pyee==12.0.0
pytest==8.3.4
pytest-check==2.4.1
pytest-html==4.1.1
pytest-metadata==3.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.17.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0